% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pullAttributes.R
\name{pullAttributes}
\alias{pullAttributes}
\title{Pull attribute metadata}
\usage{
pullAttributes(eml_filename, verbose = FALSE)
}
\arguments{
\item{eml_filename}{a set of charatures identifying the EML file downloaded from DataOne}

\item{verbose}{TRUE/FALSE to print out a whole bunch of useful(?) statements for debugging mostly}
}
\value{
a dataframe with the dataTable index, column names, descriptions, and units
}
\description{
From the EML provided by DataOne, we pull the column names, long definitions, and units.
}
