% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readBrown2013.R
\name{readBrown2013}
\alias{readBrown2013}
\title{Read in Brown 2013 (soil carbon content in Yukon River Basin Fire and Permafrost Study)}
\usage{
readBrown2013(dataDir = "data/Brown2013")
}
\arguments{
\item{dataDir}{string that specifies the data directory}
}
\value{
a list that contains the tabular dataset, a tabular version of the meta-data, the file names of the local data copies, a list of study information (abstract, copy rights, method notes)
}
\description{
Reads in data from Dana Rachel Brown and Bonanza Creek LTER. 2013. Yukon River Basin Fire and Permafrost Study: Soil physical and chemical characteristics by soil layer and year (2009-2012). LTER Network Member Node. https://pasta.lternet.edu/package/metadata/eml/knb-lter-bnz/549/4.
 This data set looks at soil carbon content from Alaska, with a tundra climate. This data set examined the effects of fire on permafrost across three soil landscapes in interior Alaska: rocky uplands, silty uplands, and sandy lowlands. This dataset includes information from each intensive study site on soil stratigraphy, texture, bulk density, moisture, water storage, pH, electrical conductivity, organic and inorganic carbon content, and ice content.
}
\examples{

}
