% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCoweeta2019A.R
\name{readCoweeta2019A}
\alias{readCoweeta2019A}
\title{Read in Coweeta 2019 Carbon and Nitrogen data}
\usage{
readCoweeta2019A(dataDir = "data/Coweeta2019C_N")
}
\arguments{
\item{dataDir}{string that specifies the data directory}
}
\value{
a list that contains the tabular dataset, a tabular version of the meta-data, the file names of the local data copies, a list of study information (abstract, copy rights, method notes)
}
\description{
Reads in data from Coweeta Long Term Ecological Research Program and Jennifer D. Knoepp. 2019. Soil percent carbon and nitrogen from 9 hillslopes sites in Macon County, North Carolina, within the Upper Little Tennessee River Basin. LTER Network Member Node. https://pasta.lternet.edu/package/metadata/eml/knb-lter-cwt/1148/13.
}
