% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGaiser2.R
\name{readGaiser2}
\alias{readGaiser2}
\title{reading in Gaiser2009(2)}
\usage{
readGaiser2(dataDir = "data/Gaiser2009(2)", verbose = FALSE)
}
\arguments{
\item{dataDir}{path to the data directory}

\item{verbose}{}
}
\value{
a list with the data, meta data, file information, and citation.
}
\description{
reading ind data collected in NE shark slough, biogeochemcial data ['https://search.dataone.org/view/https://pasta.lternet.edu/package/metadata/eml/knb-lter-fce/1060/8']
}
