% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGroffman2020.R
\name{readGroffman2020}
\alias{readGroffman2020}
\title{Read in Groffman 2012 (microbial biomass, C & N cycling, soil)}
\usage{
readGroffman2020(dataDir = "data/Groffman2012", verbose = FALSE)
}
\arguments{
\item{dataDir}{string that specifies the data directory}

\item{verbose}{a boolean flag that turns on lots of comments}
}
\value{
a list that contains the tabular dataset, a tabular version of the meta-data, the file names of the local data copies, a list of study information (abstract, copy rights, method notes)
}
\description{
Reads in data from Peter M Groffman, Lisa Martel, and Jorge Duran. 2020. Hubbard Brook Experimental Forest: microbial biomass and activity at climate gradient plots, 2010-2012. LTER Network Member Node. https://pasta.lternet.edu/package/metadata/eml/knb-lter-hbr/197/4.
}
\details{
These data were gathered as part of the Hubbard Brook Ecosystem Study (HBES). The HBES is a collaborative effort at the Hubbard Brook Experimental Forest, which is operated and maintained by the USDA Forest Service, Northern Research Station.
}
