% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readHoward2020.R
\name{readHoward2020}
\alias{readHoward2020}
\title{Read in Howard 2020 (soil carbon content in Florida Bay and Brazil)}
\usage{
readHoward2020(dataDir = "data/Howard2020")
}
\arguments{
\item{dataDir}{string that specifies the data directory}
}
\value{
a list that contains the tabular dataset, a tabular version of the meta-data, the file names of the local data copies, a list of study information (abstract, copy rights, method notes)
}
\description{
Reads in data from Jason L Howard and James W Fourqurean. 2020. Organic and inorganic data for soil cores from Brazil and Florida Bay seagrasses to support Howard et al 2018, CO2 released by carbonate sediment production in some coastal areas may offset the benefits of seagrass “Blue Carbon” storage, Limnology and Oceanography, DOI: 10.1002/lno.10621. LTER Network Member Node. https://pasta.lternet.edu/package/metadata/eml/knb-lter-fce/1228/2. 
 This data set looks at soil carbon content in Brazil, a country with a tropical climate.This data set measures soils' organic and inorganic carbon from Florida Bay and Brazilian seagrass meadows for the top 1 m of soil. Instrumental analyses and loss on ignition at 500C were used to measure C content of downcore slices.
}
\examples{

}
