% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readKelley2017.R
\name{readKelley2019}
\alias{readKelley2019}
\title{Read in Kelley 2019 (thaw index of permafrost)}
\usage{
readKelley2019(dataDir = "data/Kelley2019", verbose = FALSE)
}
\arguments{
\item{dataDir}{string that specifies the data directory}
}
\value{
a list that contains the tabular dataset, a tabular version of the meta-data, the file names of the local data copies, a list of study information (abstract, copy rights, method notes)
}
\description{
Reads in data from Allison K Kelley, Elaine F. Pegoraro, Marguerite Mauritz, Jack A. Hutchings, Susan M.N. Natali, et al. 2019. Eight Mile Lake Research Watershed, Thaw Gradient: Seasonal thaw depth 2004-2019. LTER Network Member Node.
}
\examples{

}
