% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMorris2020A.R
\name{readMorris2020A}
\alias{readMorris2020A}
\title{Read in Moriss 2020a (Porewater)}
\usage{
readMorris2020A(dataDir = "data/Morris2020Porewater", verbose = FALSE)
}
\arguments{
\item{dataDir}{string that specifies the data directory}

\item{verbose}{a boolean flag that turns on lots of comments}
}
\value{
a list that contains the tabular dataset, a tabular version of the meta-data, the file names of the local data copies, a list of study information (abstract, copy rights, method notes)
}
\description{
downloaded from https://search.dataone.org/view/https://pasta.lternet.edu/package/metadata/eml/edi/136/5
  
  ABSTRACT:Porewater nutrient concentrations were measured as a component of a long-term project seeking to understand how salt marsh primary production and sediment chemistry respond to anthropogenic (e.g. eutrophication) and natural (e.g. sea-level rise) environmental change. Feedbacks between plants, sediments, nutrients and flooding were investigated with particular attention to mechanisms that keep marshes in equilibrium with sea level. Other data collected as part of the project include aboveground macrophyte biomass, plant density, marsh surface elevation and annual above ground primary productivity. These data have been used to develop the Marsh Equilibrium Model, an important tool for coastal resource managers.
}
\details{
Sampling occurred at Spartina alterniflora-dominated salt marsh sites in North Inlet, a relatively pristine estuary near Georgetown, SC on the SE coast of the United States. North Inlet is a tidally-dominated, bar-built estuary, with a semi-diurnal mixed tide and a tidal range of 1.4m. The 25-km2 estuary is comprised of about 20.5 km2 of intertidal salt marsh and mudflats, and 4.5 km2 of open water. Sampling began at two locations in December 1993, and at three additional locations in January 1994. Sampling occurred approximately monthly through 2019. Sampling occured at a sixth location from 2006 to 2010. The site was a dieback site that had recovered by 2010. At the other sites, the study is on-going.

Porewater was collected at multiple depths from diffusion samplers and was analyzed for sulfide, chloride, ammonium, phosphate , and iron concentrations. There are five sampling locations at three sites. Two locations are in the low marsh; three locations are in the high marsh. One high marsh location had control sampling plots in addition to plots fertilized with nitrogen and phosphorus.
}
