% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readNippert2019.R
\name{readNippert2019}
\alias{readNippert2019}
\title{Reads in Nippert 2019 (soil moisture with depth)}
\usage{
readNippert2019(dataDir = "data/Nippert2019", verbose = FALSE)
}
\arguments{
\item{dataDir}{string that specifies data directory}

\item{verbose}{boolean flag to print out useful debugging statements}
}
\value{
A list that contains information including abstract, license, methods, data tables, and meta data for the data set
}
\description{
Reads in data from Jesse Nippert. 2019. ASM01 Soil water content measured by neutron probe at Konza Prairie. LTER Network Member Node. https://pasta.lternet.edu/package/metadata/eml/knb-lter-knz/11/13.
}
