% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readRice2018.R
\name{readRice2018}
\alias{readRice2018}
\title{Read in Rice 2018}
\usage{
readRice2018(dataDir = "data/Rice2018")
}
\arguments{
\item{dataDir}{string that specifies the data directory}
}
\value{
a list that contains the tabular dataset, a tabular version of the meta-data, the file names of the local data copies, a list of study information (abstract, copy rights, method notes)
}
\description{
Reads in data from Charles Rice. 2018. OMB01 Microbial biomass in the Belowground Plot Experiment at Konza Prairie (1989-1999). LTER Network Member Node. https://pasta.lternet.edu/package/metadata/eml/knb-lter-knz/53/7.
 
 ABSTRACT: The purpose of this data set is to observe long-term variations in microbial biomass in belowground plots at Konza Prairie. These effects are due to annual burning, mowing, and nitrogen and phosphorus fertilization.
}
