% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSuding.R
\name{readSuding}
\alias{readSuding}
\title{Extended Growing Season}
\usage{
readSuding(dataDir = "data/Suding2019", verbose = FALSE)
}
\arguments{
\item{dataDir}{string that specifies the data directory}

\item{verbose}{a boolean flag that turns on lots of comments}
}
\value{
a list that contains the tabular dataset, a tabular version of the meta-data, the file names of the local data copies, a list of study information (abstract, copy rights, method notes)
}
\description{
Read in data from Katharine N Suding, Jane G Smith, and Niwot Ridge LTER. 2020. Black sand extended growing season experiment aboveground net primary productivity, from 2019 to ongoing, yearly. LTER Network Member Node. https://pasta.lternet.edu/package/metadata/eml/knb-lter-nwt/265/1.
}
