% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readWall2014.R
\name{readWall2014}
\alias{readWall2014}
\title{Read in Wall 2014 (soil carbon content in McMurdo Dry Valleys)}
\usage{
readWall2014(dataDir = "data/Wall2014")
}
\arguments{
\item{dataDir}{string that specifies the data directory}
}
\value{
a list that contains the tabular dataset, a tabular version of the meta-data, the file names of the local data copies, a list of study information (abstract, copy rights, method notes)
}
\description{
Reads in data from Diana Wall and Ross Virginia. 2014. McMurdo Dry Valleys Soil Elevational Transect Experiment. LTER Network Member Node. https://pasta.lternet.edu/package/metadata/eml/knb-lter-mcm/4003/5. 
 looks at soil carbon content in Antarctica, a continent with tundra climate. This data set studied the number of soil organisms (nematodes, rotifers and tardigrades), divided by species, sex and maturity was monitored at 3 elevations, initially in Taylor Valley (1993) then Garwood and Miers Valleys (2012). In this study, the organic and inorganic carbon content was also measured of the soils.
}
\examples{
 
}
